feat(components): extract CompositeEditor & EmptyWarning Components - extract into 2 separate monorepo packages so it can be used in other frameworks like Aurelia-Slickgrid - keep full unit test coverage on both components
