+++ packages/composite-editor-component/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@slickgrid-universal/text-export",
-  "version": "0.4.2",
-  "description": "Export to Text File (csv/txt) Service.",
+  "name": "@slickgrid-universal/composite-editor-component",
+  "version": "0.4.0",
+  "description": "Slick Composite Editor Component - Vanilla Implementation of a Composite Editor Modal Window Component",
   "browser": "src/index.ts",
   "main": "dist/commonjs/index.js",
   "module": "dist/es2015/index.js",
@@ -9,9 +9,9 @@
   "publishConfig": {
     "access": "public"
   },
-  "files": [
-    "/dist"
-  ],
+  "directories": {
+    "src": "src"
+  },
   "scripts": {
     "build": "cross-env tsc --build",
     "build:watch": "cross-env tsc --incremental --watch",
@@ -38,10 +38,10 @@
   ],
   "dependencies": {
     "@slickgrid-universal/common": "*",
-    "text-encoding-utf-8": "^1.0.2"
+    "dompurify": "^2.2.2"
   },
   "devDependencies": {
-    "@types/text-encoding-utf-8": "^1.0.1",
+    "@types/webpack": "^4.41.25",
     "cross-env": "^7.0.2",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2"
