+++ packages/composite-editor-component/src/.npmignore
@@ -1,0 +1,2 @@
+index.ts
+**/*.*
