+++ packages/composite-editor-component/src/index.ts
@@ -1,0 +1,1 @@
+export * from './slick-composite-editor.component';
