+++ packages/composite-editor-component/src/slick-composite-editor.component.ts
