+++ packages/composite-editor-component/tsconfig.json
@@ -1,0 +1,41 @@
+{
+  "extends": "../tsconfig-build.json",
+  "compileOnSave": false,
+  "compilerOptions": {
+    "rootDir": "src",
+    "declarationDir": "dist/es2015",
+    "outDir": "dist/es2015",
+    "target": "es2017",
+    "module": "es2015",
+    "sourceMap": true,
+    "noImplicitReturns": true,
+    "lib": [
+      "es2020",
+      "dom"
+    ],
+    "types": [
+      "moment",
+      "jquery",
+      "node"
+    ],
+    "typeRoots": [
+      "node_modules/@types",
+      "src/typings"
+    ]
+  },
+  "exclude": [
+    "cypress",
+    "dist",
+    "node_modules",
+    "**/*.spec.ts"
+  ],
+  "filesGlob": [
+    "./src/**/*.ts",
+    "./test/**/*.ts",
+    "./custom_typings/**/*.d.ts"
+  ],
+  "include": [
+    "src/**/*.ts",
+    "src/typings/**/*.ts"
+  ]
+}
