+++ packages/empty-warning-component/tsconfig.build.json
@@ -1,0 +1,40 @@
+{
+  "compilerOptions": {
+    "module": "es2020",
+    "moduleResolution": "node",
+    "target": "es2015",
+    "lib": [
+      "es2020",
+      "dom"
+    ],
+    "typeRoots": [
+      "../typings",
+      "../../node_modules/@types"
+    ],
+    "outDir": "dist/amd",
+    "noImplicitAny": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "noImplicitReturns": true,
+    "skipLibCheck": true,
+    "strictNullChecks": true,
+    "declaration": true,
+    "forceConsistentCasingInFileNames": true,
+    "experimentalDecorators": true,
+    "noEmitHelpers": false,
+    "stripInternal": true,
+    "sourceMap": true
+  },
+  "exclude": [
+    ".vscode",
+    "src/examples",
+    "src/resources",
+    "test",
+    "**/*.spec.ts"
+  ],
+  "include": [
+    "../typings",
+    "**/*"
+  ]
+}
