+++ packages/excel-export/README.md
@@ -14,7 +14,7 @@
 
 Simple Export to Excel Service that allows to exporting as ".xls" or ".xlsx".
 
-### Dependencies
+### External Dependencies
 This package requires [excel-builder-webpacker](https://www.npmjs.com/package/excel-builder-webpacker) which itself requires [jszip](https://www.npmjs.com/package/jszip) and [lodash](https://www.npmjs.com/package/lodash), the later not being a small lib, so make sure that you are fine with the bundle size. For our use case, the extra bundle size is totally worth the feature.
 
 ### Installation
