+++ packages/excel-export/package.json
@@ -37,7 +37,7 @@
     "not dead"
   ],
   "dependencies": {
-    "@slickgrid-universal/common": "^0.4.0",
+    "@slickgrid-universal/common": "*",
     "excel-builder-webpacker": "^2.0.1",
     "moment-mini": "^2.24.0"
   },
