+++ packages/odata/README.md
@@ -14,7 +14,7 @@
 
 OData Service to sync a grid with an OData backend server, the service will consider any Filter/Sort and automatically build the necessary OData query string that is sent to your OData backend server.
 
-### Dependencies
+### External Dependencies
 No external dependency
 
 ### Installation
