+++ packages/text-export/README.md
@@ -18,7 +18,7 @@ There are a couple of reasons to use this package (instead of the `excel-export`
 - if you want to export to a text file with any type of separator (tab, colon, semicolon, comma)
 - if you have a very large dataset, this export consumes a lot less memory compare to the `excel-export`
 
-### Dependencies
+### External Dependencies
 This package requires [text-encoding-utf-8](https://www.npmjs.com/package/text-encoding-utf-8) which is use to ensure proper UTF-8 encoding, even emoji will be exported without issues.
 
 ### Installation
